python/requirements.txt
tdl-client-python==0.10.1